Add scanning for options
This is required for sanity checking.  It also depends on functionality
that isn't in bpm yet, but it should be (it will at least be in dports,
but obviously it must in ports for FreeBSD users as well).

The "Packages" child package has been started but it will change.
This is basically a checkpoint.