Populate var folder and add suppression option to execute()
While here, also fix bugs introduced in previous commit.
The "mtree" progrom emits output to stdout no matter what.  It can't
be suppressed.  To counter, add a "suppress" mode (off by default)
to the execute() procedure that will direct the output to a temporary
file which is immediately deleted afterward.