Bite the bullet and open/close logs repeatedly
I could not figure out a way to limit the build log opening and
closing to one time each.  The spawning code needs a file descriptor
and I failed to find a clean way to extract it from the File_Type.
Having both File_Type and an open file descriptor will not throw
an error, but neither will the text write to the log.

Since only one file handler can be active at a time, the File_Type
handler has to be closed before every phase spawn, and reopened
when it's done (to create the phase header/footers).