Implement through package
It was discovered that the passwd stuff had to be regenerated, and
similarly the group file had to be present as well.  I didn't want
to rely on /usr/src being present and we have to support live system
building (a major goal of the project) so both passwd and group come
from the live system but with only the system users and groups
present (e.g. clean system).

It was noted that /packages needs to be a r/w mount as well.
Synth is building packages now but I'm seeing bogus messages like
"required shared library libm.so.4 not found" during packaging.  I
think DragonFly poudriere does this as well so I need to check
FreeBSD poudriere for the key.