Illustrate how to handle pkg(8)
Pkg(8) is a prerequisite for every port (except itself).  Rather than add
it as a dependency for 25,000+ ports which only has to be removed
immediately, we'll just always build it first.  While the former is
cleaner, it also needs well over a minute to accomplish on poudriere.  It
wouldn't take nearly as long on synth, but it would cause a lot of
alloc/dealloc of container memory.