Speed up (significantly?) the library scanning
I forgot that /usr/bin/file can scan the entire file trying to figure out
if it's binary or not.  It is NOT suitable for use for the library
scanning.  I decided to just use objdump by itself, only "grepping" if
the return code was 0 (this unfortunately allows static libraries to
fall through, but it should filter out text files).  The final result
should be faster overall.