Add ability to disable build logs
I turned off the build logs completely, yet standard was still clobbered.
I previously turned off the pipes (which is also turned off when the
logs are turned off) and stdout was still killed.  That leaves me to
believe the problem is with OSL.spawn....

I've got no idea how to fix this.