FIXED IT!
The issue is that all the concurrent spawnings were emitting to stdout,
the stream was shared.  So when the parent was emitting text, it was
getting sucked up into any number of spawns that were redirecting
standard out to a file.

The only solution I could come up with was to create an external
help to spawn a single command to file.  I called this helper
synthexec.

example of use:
spawn /tmp/mylog /usr/local/libexec/synthexec /usr/local/bin/pkg -vv

the output of pkg -vv is pushed to /tmp/mylog.  No crossing over.
Everything is working now AFAICT.