Fix synth.ini creation issue, move to etc/synth/
When testing the move of synth.ini to etc/synth/, I discovered the
ini creation had been broken by the last set of commits to parameters.
Morever, the mount check in the main synth file depends on the
configuration, so it needs to be loaded first.

To support the move, add a new "mkdirp" function to create
directories as necessary.