Add synthexec detection, LANG=C, refactor phases
I lost a bunch of time trying to figure out why synth suddenly was
failing before I failing tracked it down to a missing synthexec.  Since
this is a critical component, check for it at launch.

In preparation for build tracking, split phase-to-string to its own
function and tightly up build package (tighter but maybe less readable)
Anyway, support storing current phase and also counting the current
lines in the log (for progress purposes)