Implement hooks system and first two hooks
The hooks system works.
In the /usr/local/etc/synth directory, it looks for three EXECUTABLE
files called run_start, run_end, and pkgbuild.

The run_start program is spawned at the beginning of a bulk build.
The run_end program is spawned when the build is finished.  Both
provide statics on the port build numbers.

The third one, pkgbuild, will fire when a pkg cycle is complete.
It has a RESULT value (env. var) set to either SUCCESS, FAILURE,
IGNORED, or SKIPPED along with other data specific to how the
build ended.  All data is passed with env var.

The pkgbuild hooks will be implemented tomorrow.