Implement pkg hooks (four, not one)
I decided to split up the pkg hook into 4 pieces so that the hook is
no run unnecessarily (this is unlike poudriere which has one hook
for all for results).  I will improve the executable check to accept
symlinks so it will be possible to symlink each hook to another.

new hooks:
hook_pkg_success
hook_pkg_failure
hook_pkg_skipped
hook_pkg_ignored

That the four ways a package build can end, and the appropriate hook
is called accordingly.