Minimally tested sanity check overhaul
This overhaul seeks to significantly improve the time to do a sanity
check on a large repository.  It does one ABI and option check.  If it
passes those, it queries the existing package and stores the result.
After all the packages have been queried, it starts marking them for
deletion, repetitively until there are no more failures.  Obviously
it uses the cached query results instead of spawning a new pkg request.
upon a clean exit from the loop, it deletes all the packages marked
for deletion.

If all is correct, it should end up with the same results as before,
but much, much faster.