Load/swap only had range 0 <=> 10
I didn't realize that "digits" limited the range of the type concretely.
Digits of 3 with a delta of 0.01 makes a range of 0.00 => 9.99.  Any
load or swap that is 10 or more fails.

Fix it by changing digits to "4".  Loads > 100 are already handled so
another exception will not happen.  It was only luck that I didn't see
swap > 10%; that would have failed as well.  This should finally fix
the full bulk run on Muscles.