Fix alignment for load > 10
The good news: The previous fortication "worked" in that the forgotten
leading space on 'Img did not crash the routine, but it did cause it
to shift over (as intended).  Now that we now "digits" is rigid, we
know we can't have longer strings than 5 after the leading space is
stripped off, so do that and remove the unneeded cases/exceptions.