Disable SIGINT handling until further notice
The SIGINT capture is propagating to the synthexec spawn and anything
that emits it (say a failed conftest) causes the spawn to get stuck
(the process just waits for something to happen).  Nothing I've tried
can work around anything spawned by Synth.

I really want this back.  I'll probably map ESCAPE key to this in the
meantime.