Rather than fail, skip installed packages when they have no port
It's possible to install a port that later gets pruned from the tree,
or install a port from another tree that doesn't exist in the main
one.  If Synth status finds such a port, it now just mentiones that
it will be left off the update list and continue (scenario arose first
with kwm, but it's a common one).