Drastically reduce escape checks during package scanning (for FreeBSD)
With the original SIGINT handler, having each task check if control-c
had been pressed was not heavy task; it was just a boolean check.  When
that got changed to checking for keypresses, well, FreeBSD choked on it.
Nobody had any issues on DragonFly, but FreeBSD would frequently (e.g
4 out of 5) fail to return on a small set of packages.  Also, there are
32 tasks that all fire at once, then most of them shut off.  I think the
burst of checks is what locked FreeBSD up (by the way, pressing Escape
repeatedly would allow FreeBSD to finish, so it was just waiting for more
keypresses.  Likely it's a bug for FreeBSD, but it's also bad form on
Synth's part.  Just move the check to occur once per second and all
taskers can read the result as a boolean once again.

While here:
  * rename passed_initial_package_scan => initial_package_scan
    It's a procedure now, not a function
  * Add an exception handler on the package option check
  * Add an exception handler on the package dependency check
  * Add an exception handler on the package abi check