For FreeBSD, add linprocfs mounts to support Linux packages
The linprocfs mount works but the first package I tried results in
a message that all requested packages are built (they aren't) so
Linux support is a work in progress.

Mounting linprocfs causes FreeBSD to autoload the Linux kernel
module, so I don't need to mess with kldload at all.