Fix repo creation (error illustrated on FreeBSD)
Rebuilding the repository was using code that used logging that assumed
there were build queues.  It broke on FreeBSD and I'm not sure why it
didn't break on DragonFly.  This was the same problem seen with scanning
packages earlier (standalone routine was created for pkg installation)

While here validate the argument for get_catport.  It should never be
invalid but I hit this.  I still need to find the original cause those.