Still the directory /usr/ports/distfiles during scanning
On FreeBSD, the default distfiles location is /usr/ports/distfiles
(brilliant) and until now, Synth has considered this a valid category.
It doesn't check category nows, only that the directory starts with
lowercase 'a' through 'z', but unfortunately 'distfiles' satisfies
that requirement.  Specificially check for this and skip if found.

Reported by: Stari Karp