Cut 45% of port scan time on FreeBSD
On a virtualbox VM, a portstree scan takes 9 minutes, 30 seconds (!).
By caching a few values, the same scan now takes 5 minutes, 15 seconds
on the same VM.  That's a reduction of nearly half.

I still need to implement the equivalent on an index so the subsequent
install commands are much faster, but this is a nice performance bump.