Implement hanging process watchdog
Change the blocking spawn to a non-blocking spawn and check the pid
with a non-blocking wait command (requires C interface unfortunately).
For each phase, a maximum amount of time with no new log output is
defined (resolution is by line since that's already known, so adding
characters to the last line will be undetectable and thus considered
not progressing).

Four times per second the pid is checked to see if the process is done.
After the maximum time for that stage passes, the line count is checked
once per minute against the value that existed at current time - maximum
time.  If they are the same, the watchdog pkills the parent process and
waits 5 seconds before returning (to allow any error messages to print
to the build log).