Make load indicator insensitive to locale on FreeBSD
On FreeBSD, but not on DragonFly, the sysctl vm.loadavg output is
sensitive to the locale, meaning the decimal separator might be a
common.  This causes a conversion failure within Synth, resulting in
a load indicator that always reads zero.

By forcing the locale to C/POSIX when invoking this command, this
should restore synth's load indicator when names locales are used.