Re-enable watchdog except when scons is used
I can't reproduce the scons issue in a separate testcase.  It always works
"in the lab" but it must be some incompatibility with the threading of the
tasks.  Several workarounds since have also failed.

With the (weak) assumption that *only* scons is an issue, this commit will
enable the watchdog unless the port uses scons.

A new package, Unix, was added to split out the related routines.  This was
mainly due to the failed workarounds, but I've keep the (unnecessary)
changes anyway.

This requires another change the /usr/ports/ports-mgmt/synth though.