Turn watchdog off again, but make it switchable
There are a few more freezes (python3* are big ones) so the watchdog is
still not ready for showtime.  Let's make it easy to turn back on by
using a constant though.  Let's also make sure it doesn't try to kill
the process tree with a bad group process id so that the watchdog really
is fully disabled.