Progress on leftovers check (testmode)
Implement the "mark filesystem" routine which is required to check
for post-stage leftovers (or removals/modifications). Unlike poudriere,
we can use md5 hashes rather than rely on size differences because we
can limit the number of files due to exclusion of read-only paths.

This is a work in progress, the test mode doesn't use this yet.