Don't prepopulate /usr/local on DragonFly
It's not done on FreeBSD (and everything is fine), and moreover doing
it on DragonFly results in failed system change checks (e.g. the port
deinstallation removes /usr/local/share/examples because the ports
framework removes empty directories automatically now)

Just remove it completely to fix the system check and reduce differences
between FreeBSD and DragonFly.