Implement file system check between configure and build (inclusive)
This modifies the existing functions to be generic so we can easily add
a new check before configure and after build.  For that check, we also
check the file modification time.  This does the "file system violation"
check fairly effectively.

As a side effect, I had to customize to more generic stages in the cycle
though.