Change graceful exit from ESCAPE key to Control-Q key combination
The use of the Escape key as the shutdown method had some issues, probably
related to terminal ANSI codes.  There was also a report that spinning the
mouse wheel caused a shutdown.  Perhaps choosing "Escape" as the new
shutdown code wasn't the best idea, so for attempt #3, let's move to the
Control-Q combination (which also requires 2 keys to be pressed and thus
pretty difficult to do it inadvertently, another criticism of "Escape")