New feature: SYNTHPROFILE environment variable (issue #19)
For scripting purposes, and to fully take advantage of Synth profiles,
let's allow the user to switch profiles temporarily on the fly.  Rather
than adding new commands to pass to Synth, let's use the environment
instead.  To use it, set SYNTHPROFILE to the name of an existing
profile.  If the profile doesn't exist, the default profile is loaded
as usual.