Reincarnate "rebuild-repository" command with new purpose (issue #15)
Bring back "rebuild-repository" (after a period measured in minutes), but
now the command just performs a sanity check on the existing packages and
rebuilds the local "Synth" repository on command.  It's most likely use
case is for scripts, but people could use it in conjunction with a series
of "just-build" commands as well.