Exclude Synth repository from external repo update command
If for some reason, the Synth repository was configured for pkg(8), but
no repository was present, every prefetch check would spew errors because
it would try to update Synth repo along with the external repository.
Make it smarter by limiting the external repo update to just a single
repository that is not Synth.  This effectively removes the unwanted
error messages and prevents it from touching local repo when it should
not.