New feature, equivalent to poudriere testport -i feature
This gives the ability to break into a port builds at one of the
following points:
  - extract completion
  - configure completion
  - build completion
  - stage completion
  - install completion
  - deinstall completion

To activate the feature, issue the "test" command with a *single*
port and have ENTERAFTER set in the environment.  If all those
conditions are satisfied, Synth will dump the user into the
build with a tcsh shell at the point requested.  The user will
issue the "exit" command to leave the shell at which point Synth
will clean up (the build doesn't continue; the run is over)

This is a troubleshooting feature for advanced users (port devs)