Fix ABI check (logic flaw)
There is still debugging code in here, intentionally.
Kris Moore discovered that running synth on system root that is a
different release that the host will result in an ABI check failure.
The reason is that uname was used to calculate ABI and that's tied
the the kernel regardless of jail.

I've changed the logic to use file(1) to read /bin/sh and determine
the ABI (and alt ABI) from that.  When it's confirmed to work on
FreeBSD as well, I'll remove the debugging code.