Implement UNAME_x environment variables (issue #31)
The UNAME_v variable should be something like "FreeBSD 11.0-SYNTH"
but I just found out that ada-util external library can't accept
arguments with spaces in them (quoting doesn't work) nor extra
spaces (e.g. "cmd A B" is okay but "cmd A  B" is not).

While here, make the build abort immediately if the the
"get environment" command for the beginning of the log fails with
a discovery error.  If it fails, every command will fail, so make
it abort immediately.