Work around logic flaw (?) in emulators/virtio-kmod
For some reason OSREL is no longer being defined when <pre> is
included and all new things like _OSRELEASE are now defined in
make.conf.  The reason is not clear and I suspect it's actually
a bug in the framework.  However, just define OSREL because we
know what it is supposed to be.  This fixes the issue, at least
where it arose on this port.