Set graceful shutdown to capital "Q"
It turns out that Control-Q only works in ncurses mode (probably because
the tty is set to raw mode then).  Leave the Control-Q handling, but also
detect when capital "Q" is pressed.  The latter is now documented, but
the former continues to work (undocumented) for those that have gotten
used to it.