Return to Control-Q
I hated the idea of both changing the graceful shutdown code again, and
the fact that the next candidate, 'Q', would echo garbage to the screen.
The echo issue was generic; it has been messing with the progress output,
etc.  So I needed a way to both disable the echo and allow Control-Q to
be used.  With two small C functions, I was able to accomplish this
switching on command.  Control-Q works always now and the unwanted echoes
have been suppressed.

While here:
  * Remove Q handling and fix the man page
  * Fix bug with status-everything where a dry-run abort would emit a
    "sorry" message.
  * Also make "everything" skip the repo build if a previous abort was
    commanded
  * move nohang_waitpid binding to unix.ads like the other bindings

Greetings, Maxwell Smart.