Fix piped command exit statuses (issue #33)
There were two inter-related problems.

The first one appears that if the pipe isn't read, the pclose return
status isn't accurate.  I discovered this accidently with a heisenbug.
As soon as I read the pipe to output the message, the exit status
changed to "good".

The second bug was the exit status was not being returned correctly.
Now if the process exits abnormally, you get that exit status, and if
it exits normally but with a non-zero value (the child status), then
that is returned.

As a new feature, if a muted piped command does result in an
unexpected non-zero status, synth will output the result of the
command (which contains the error message) and that will be followed
by an unhandled exception that contains the command text.

This would havd revealed the problem with mtree in issue #33.