Major enhancement: Improve repository building times
The robust method of building the local repository takes 4-5 minutes on
DragonFly using a core i5 and a solid-state drive.  Using FreeBSD, a lesser
CPU and a mechanical disk drive could result in 20 minute repository
generation times.  The original method did a full ports tree, then it
removed all packages that weren't covered in that scan.  Then the
remaining packages were validated.

Now the packages are prescanned to determine the port origin.  If the
origin doesn't exist, the package is removed.  If it does, the port is
scanned to check if the package name is current.  This is all done in
parallel.  If both checks are passed, the port origin is put in a list
for a recursive scan, which unfortunately has to be done serially.  Once
the scan is complete, the remaining packages are validated and finally
the local repository is (re)generated.

This approach results in signficantly improved repository build times
(e.g. 63 seconds for 1085 packages in DragonFly)