Avoid exception caused by detached terminal
Running a command like "synth force" in cases where there's no attached
terminal resulted in a Text_IO exception (END_ERROR) after the build
when the questions were posed to user.

Two common cases of this is scheduling a cron job (with force) or piping
arguments to synth via xargs.

To fix, check for an attached terminal.  If there is none, just assume
the answer to the question is "false" to conclude the build.