Enhancement: Show version differences on status command
Per request, the "status" commands now provides more information about
each port to be rebuilt.

If no package exists in the packages directory, a symbol of "N" (for
"New") will be show on the display and "(New <version>)" will be shown
on the log (where "<version>" is the current port version).

If a package already exists and it's the correct version, but it's
slated to be rebuilt, a symbol of "R" will be shown on the display and
"(Rebuild <version>)" will be added to the log.

The remaining case is a package exists but it's the wrong version.  In
this case, the display will show a "U" for upgrade and after the port
origin, the old version and the new version will also be shown on the
display.  On the log "(<old-version> => <version>)" will be shown.

Per request by Marc Pearsonti