Fix long-standing purge-distfiles bug (issue #41)
The container that held the entire list of distfiles had a range
that was capped by the number of existing ports, and it was counting
each distfile separately (incrementing by 1).  Well, there are more
than one distfile per port and the container's element value busted
out of the range.

The fix was set the container element value to the index of the
*first* port that lists it.  The value isn't used anyway, so it
doesn't really matter; it's the key that we are looking at.

Buggy behavior noticed by: Bill Richter