Support signing repository with generated RSA keys (issue #39)
If these files exist in /usr/local/etc/synth/:
  [profile]-public.key
  [profile]-private.key
and the private key file permissions mode is 400, the generated
repositories will be signed with it.  Additionally, the pkg repo
configuration file, 00_synth.conf, will be modified to indicate
the PUBKEY signature and given the path to the public key.

The man page will be updated in an upcoming commit.