Document local RSA key signing (issue #39)
Changes to the man page include two new entries in FILES section for each
of the two RSA keys, and adds a new EXAMPLE section.  The latter shows
how to generate the RSA keys using the expected names.

While here, tweak the purge-distfiles command description to avoid any
ambiguity or assumption that it works the same as portmaster (it doesn't).

It will not be difficult to add external server signing support in a
similar fashion (file containing external command and second file
containing generated fingerprint for clients)