Fix "install" command regression (issue #42)
I reused the portlist variable for the fast method of creating the
repository.  This caused a regression when the Synth is told to install
the specified ports immediately afterwards.  The original specification
was replaced by a full list of every port in the repository, so
everything gets installed.  It's a pretty serious bug actually.

The solution was to duplicate the original specification list and use
that list for any post-repo installation commands.