Prevent use when cwd is inside ports directory
Synth will malfunction on the "store_origins" function if it's launched
from inside the ports directory.  I had to move up the launch check
routine to happen before this point.

Note that the check will fail if synth's ports directory is configured
to a symbolic link.  It doesn't do a "realpath" check on it (although it
probably should; maybe some day)