Enhancement: Add "realpath" functionality (for configure)
This adds a generic interface to libc's realpath function.  It is
used for checking the input of directories on synth configure.  If
someone provides a symbolic link or a path with relative components,
Synth will automatically convert it to the fully resolved "real" path.
This is to ensure the launch checks always work, among other things.