Fix potential prefetching error
The "pkg rquery" command appears to be sensitive to the order of the
arguments.  The "-r" option must come first, otherwise at least some
operations are down on all the registered repositories.  If one is
misconfigured, e.g. Synth, then "pkg rquery" is preceded by error
messages, thus breaking all the ABI/options/deps checks.

This should fix the issue which was reported by "Synth Meat" in the
forums.