Change pre-fetch behavior and handle fetch failure
Previously, "synth test" or "synth force" command combined with prefetching
option enabled, the listed ports would fetch and not build.  However,
"force" and "test" imply the user really wants the package(s) built, so
this change omits the listed ports from fetching.  So "force" and "test"
will pre-delete the package and not fetch a new one (regardless of the
pre-fetch setting) so they are always built.

Moreover, previously after a fetch failure, a question about rebuilding
the repository would be asked with "force" command.  Now if a fetch
failure occurs, Synth just ends.