Enhancement: Add LOCK feature to test mode
if LOCK is defined to any value in the environment when "synth test"
is issues, the builder's /usr/local will be remounted as read-only
prior to the configure phase and remounted again with read/write
permissions when the build phase is complete.  This is a diagnostic
tool to help pinpoint when ports access this area illegally during
these phases.