Enhancement: Alert user when pkg(8) CONSERVATIVE_UPGRADE is on
If the user runs the "synth rebuild-repository" command and the system's
pkg(8) configuration has CONSERVATIVE_UPGRADE set to "true", then Synth
will now display a notice when the command is complete.

If the user intends to run "pkg upgrade" to update the system from the
newly-built repository, they probably don't want this setting on.  Most
people are likely unaware of it, so it serves as education.