test mode: ignore /var/db/fontconfig changes always
The /var/db/fontconfig directory was only being ignored post-build
to change changes during the install, but according to the doxygen
port, /var/db/fontconfig can be modified between configuration and
the end of building.

This appears legitimate to me, so move the /var/db/fontconfig
exception from post-build to all stages.