Provide instructions on how to deal with bad OPSYS definition
It's not really possible to handle a bad opsys definition automatically,
so the next best thing is to explain the problem and give the solution
to the user.

In this case, we tell the user that the opsys is a bad value, we tell
them the two possible good values, and also to remove the offending line
in synth.ini manually.  The next command wil regenerate the profile if
everything (ports, pkg) is in good shape.