Simplify reaping
The PROC_REAP_RELEASE isn't necessary and seems to cause errors when
called on FreeBSD so remove it.  Also, check globally that the given pid
is greater than zero (it should always be)

The dragonfly implementation was wrong, we have to iterate through
PROC_REAP_STATUS until it returns no pid.  Take the pids it does return
an kill the process and wait for it.  Finally wait for the reaper_pid
too.