Double run_depends time limit again (10 minutes to 20 minutes)
I don't know what math/octave-forge is doing, but under contention, it
was trying to load some octave thing in run-depends phase and needed
more than 10 minutes to do it.  By itself the entire octave-forge built
in 16 minutes  but takes as long as 55 in a full bulk build on muscles.
It seems that contention plays a part here.