filter master.passwd too (don't copy it straight)
Matt Dillon discovered that the passwd file has to be filtered out to
match the tailored group file.  The system assumes that if the user
exist, it's listed group does as well.  Remove the straight copy and
limit the master.passwd entries to those found in the starting version
(a combination of both FreeBSD and DragonFly)