Introduce load-based timeout multiplier
The idea is that the 5-minute load average is checked at the beginning
of each phase and it helps determine the timeout interval.  It rounds
the value of the load divided by the number of cpus to the nearest tenth
(1.0 is the floor) and that value is multiplied by the base timeout
value.

For example, a four-core machine with an average load of 6.0 will have
an average load of 1.5.   A phase with a 10-minute watchdog interval
base will have a timeout of 10x1.5 = 15 minutes.