Tweak limits of check-plist, build (issue #57 )
The 3-minute limit on check-plist can be too tight on a heavily loaded
machine when the package it's building has thousands (if not tens of
thousands) of files that need to be checked.  Bump that from 3 to 10
minutes.

There also seems to be edge cases involving very old machines where 20
minutes (base) is allowing the watchdog to trigger prematurely.  Add 25%
more time to the base build phase limit (from 20 to 25 minutes)