Update log counters in text mode to (issue #58 )
The watchdog relies on the log counters which were only being updated in
the ncurses display mode.  This change updates the builder log counters
every 30 seconds in the text mode.

I was able to reuse the 30 second refresh counter in text mode, but if
we want resolution lower than this, I'd have to create a dedicated
counter.  Since the ncurses mode updates the logs every second, that
means up to 29 seconds are effectively shaved off the watch dogs, but
they are generous enough already that this shouldn't be of practical
consequence.