Rework display to handle resizing better
1) If the display width is too short, the lines won't wrap anymore
2) certain elements like the elapse timer will not display in the wrong
place anymore (this was due to a failing move command)
3) the static areas are now redrawn if the window area expands to expose
them (before they would remain blank forever)
4) The blank areas are refreshed every 4 seconds instead of every 30
seconds [A]

[A] This has the side effect of the text mode logs being checked every 4
seconds instead of every 30 seconds as well.  The same variable is
reused.  It's okay.

The "Choose_Pen" routine is no longer required so it's been removed.
The "Scrawl" prototype was changed to have relocate parameters and now
it handles relocation as well.   This was required for error handling,
but it simplifies things as well.