Fix minimum history height (and add maximum)
I'm not sure what was going on here, but it appears the minimum history
window height wasn't coded as desired.  Update it to limit the range to
a minimum of 10 rows with an upper limit of 50 to match the history
cycle definition.  This allows the sanity check later down to be
removed.