WIP: set up web reporting
Primarily this adds a function to create summary.json file after copying
all the web assets from share/synth into the log directory's
subdirectory "report".

Some existing functions had to be relocated / exposed for refactoring
purposes.
The "remaining" name in the report was also renamed to "remains" for the
sake of name length.
This is untested, uploading to github for purposes of testing it.