Concept rework: increase flavor index
The all_ports data is not available for the upgrade_system_everything.
In fact, the pkg query is supposed to define what to scan.  Since we
can't use it, but we need the pkgname information that it contains, we
have to add it to the flavor index.

If this solves the issue, I will make the rebuild repository routine use
the same technique to keep them in sync (simpler, proven)