fix issue#199
There are 2 directories where if they fail to unmount, we've got significant problems:
1) distfiles directory
2) packages directory

The parent directory to their mountpoints can have "rm -rf" applied to it,
and if they are still mounted, it wipes out everything in distfiles and/or packages.

The --force flag was original left off because of potential panics of forcibly
removing mounts that thought they were busy but this seems to be safer
than it was in the past.